Fix Win32 build
[ ] May break build
[ ] Breaks existing applications
[x] Bugfix
[ ] New Feature
[ ] Cleanup
[ ] Optimization
[ ] Documentation

+++ lib/base/spinLock.h
@@ -86,7 +86,7 @@ namespace base
         bool isSet() { memoryBarrier(); return ( _set == 1 ); }
 
     private:
-        int32_t _set;
+        long _set;
     };
 }
